[PATCH] deal with tags for good
keep a global hash of tags, do not allocate a new structure every
time we meet a tag in some rcs file, add a reference to commit into
set for that tag instead.  Set is kept as linked list of blocks.
That has an extra benefit of eliminating rev_find_tag() nonsense
(the worst timesink remaining in the first and second stages).

Signed-off-by: Al Viro <viro@zeniv.linux.org.uk>