[PATCH] crude workaround for FreeBSD RELENG_2_1_0 braindamage
Some idiot had tried to retcon RELENG_2_1_0 after a fsckup
and ended up making it an alias for HEAD on some files (look
for old_RELENG_2_1_0:1 in rcs files and you'll see; e.g.
usr.sbin/sysinstall has a bunch of those).  The best we
can do is to act as if branch is absent in such files...

Signed-off-by: Al Viro <viro@zeniv.linux.org.uk>