Check commit authors. Add --commit-time-window option.
Improve commit comparisons so that commits with different authors are
assumed to be different.

Add --commit-time-window to allow people to change the time interval allowed
between commits that could still be considered the same. In particular, a
value of '0' will make all commits be different that don't have the same
commitid.