Clean up out-of-order vendor branches (thanks, LaMont). Avoid master loops.
Vendor branches will occasionally get recorded with out-of-order dates,
this confuses the vendor branch merging code. Reorder the vendor branch to
preserve monotonicity. Also, allow the 'master' to end on a vendor branch
revision by not checking for branch merging for the master branch. This
prevents a loop along the master branch.